Add LinkHelper class
The LinkHelper is designed to be used by any containerservice template (or other code) that needs to link to a separate stack so we provide this as a global helper available to all templates automatically.

Parameter values are automatically escaped so that the generated URL created by the LinkHelper are valid and do not create any problems when opened within a browser.